CmdCheck now refers to populate_table and adds a column for interrupt chances. This will be the last addition to the filter_and_modify_arts branch, so once I make any recommended changes based on comments, I plan to merge it and move on. I had to 1) move the table-related functions (setup_table, populate_table, abbreviate_effects, and modify_ap_string) from utilities.py to a new file, tables.py, and 2) pass the caller object to populate_table, which I tried to avoid doing in the past. The reason is that CmdCheck calls populate_table when it is displaying a character's chances of interrupting a single specific incoming attack with each of the interrupting character's available arts and normals, so "interrupt_chance_calc" must be called from combat_functions.py in populate_table itself. Interrupt_chance_calc needs the caller and combat_functions.py refers to utilities.py, so to avoid a circular reference to combat_functions, I made a new tables.py file. Ugen suggested an alternative: create an object that contains the relevant caller information but is not the caller and send that to various combat functions. My concern right now is that I'm still developing aspects and that I'm not sure what all is ultimately going to go INTO such an object, but I think that would be a good change once I'm done with this "combat system 2.0" pass, so I'm not unnecessarily passing around the whole character object everywhere.
